+++ apps/common-app/runtime-tests/worklets/suites.ts
@@ -45,6 +45,14 @@ export const WORKLETS_TEST_SUITES: RuntimeTestSuite[] = [
     disabled: !isBundleModeEnabled() || !__DEV__,
   },
   {
+    testSuiteName: 'networking',
+    importTest: () => {
+      require('./tests/networking/fetchApi.test');
+      require('./tests/networking/fetch.test');
+    },
+    disabled: !isBundleModeEnabled(),
+  },
+  {
     testSuiteName: 'run loop',
     importTest: () => {
       require('./tests/runLoop/requestAnimationFrame.test');