+++ packages/react-native-worklets/Common/cpp/worklets/WorkletRuntime/WorkletRuntime.h
@@ -27,6 +27,8 @@ namespace worklets {
 using namespace facebook;
 using namespace react;
 
+class Networking;
+
 template <typename TCallable>
 concept RuntimeCallable = std::is_same_v<std::remove_cvref_t<TCallable>, jsi::Function> ||
     std::is_same_v<std::remove_cvref_t<TCallable>, std::shared_ptr<SerializableWorklet>>;
@@ -245,7 +247,8 @@ class WorkletRuntime : public jsi::HostObject, public std::enable_shared_from_th
       const std::string &name,
       const std::shared_ptr<AsyncQueue> &queue = nullptr,
       bool enableEventLoop = true,
-      bool enableLocking = true);
+      bool enableLocking = true,
+      bool enableNetworking = true);
 
   void init(const std::shared_ptr<JSIWorkletsModuleProxy> &jsiWorkletsModuleProxy);
 
@@ -420,7 +423,8 @@ class WorkletRuntime : public jsi::HostObject, public std::enable_shared_from_th
   void bundleModeInit(
       const std::shared_ptr<JSScheduler> &jsScheduler,
       const std::shared_ptr<const ScriptBuffer> &script,
-      const std::string &sourceUrl);
+      const std::string &sourceUrl,
+      const std::shared_ptr<Networking> &networking);
 
   void legacyModeInit(const std::shared_ptr<UnpackerLoader> &unpackerLoader);
 
@@ -433,6 +437,7 @@ class WorkletRuntime : public jsi::HostObject, public std::enable_shared_from_th
 
   const RuntimeData::RuntimeId runtimeId_;
   const bool enableLocking_;
+  const bool enableNetworking_;
   const std::shared_ptr<std::recursive_mutex> runtimeMutex_;
   const bool microtaskQueueEnabled_;
   const std::shared_ptr<jsi::Runtime> runtime_;